flytt funksjoner knyttet til restYtelsePerioder inn i RestTilkjentYtelse.kt og lag tilhørende testfil
